Move `Protocol` and `Circuit` classes
In preparation for TH quotes for instances of classes, the modules are
rearranged to prevent import loops when adding the TH quotes.
