+++ clash-protocols.cabal
@@ -190,6 +190,7 @@ library
   other-modules:
     Data.Bifunctor.Extra
     Paths_clash_protocols
+    Protocols.Internal.Classes
 
   default-language: Haskell2010
